Use the Spring Initializr
I rewrote the document and edited the code (mostly changing package names) to use the Spring Initializr

I also upgraded the build files to Spring Boot 2.2.0, which necessitated changing the tests to use JUnit 5 and replacing the gradle wrappers.
