+++ complete/src/main/java/com/example/springboot/HelloController.java
@@ -1,14 +1,14 @@
-package hello;
+package com.example.springboot;
 
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @RestController
 public class HelloController {
-	
+
 	@RequestMapping("/")
 	public String index() {
 		return "Greetings from Spring Boot!";
 	}
-	
+
 }
