+++ initial/src/main/java/com/example/springboot/Application.java
@@ -1,4 +1,4 @@
-package hello;
+package com.example.springboot;
 
 import java.util.Arrays;
 
@@ -8,17 +8,17 @@ import org.springframework.context.ApplicationContext;
 
 @SpringBootApplication
 public class Application {
-	
+
 	public static void main(String[] args) {
 		ApplicationContext ctx = SpringApplication.run(Application.class, args);
-		
+
 		System.out.println("Let's inspect the beans provided by Spring Boot:");
-		
+
 		String[] beanNames = ctx.getBeanDefinitionNames();
 		Arrays.sort(beanNames);
 		for (String beanName : beanNames) {
 			System.out.println(beanName);
 		}
 	}
-
+	
 }
